refactor: move utils cobra to command package
