+++ cmd/cli/kubectl-kyverno/command/docs.go
@@ -1,4 +1,4 @@
-package cobra
+package command
 
 import "strings"
