+++ cmd/cli/kubectl-kyverno/command/docs_test.go
@@ -1,4 +1,4 @@
-package cobra
+package command
 
 import (
 	"testing"
