+++ cmd/cli/kubectl-kyverno/commands/command.go
@@ -1,6 +1,7 @@
 package commands
 
 import (
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/apply"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/create"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/docs"
@@ -10,15 +11,14 @@ import (
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/test"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/version"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/experimental"
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
 	"github.com/spf13/cobra"
 )
 
 func RootCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "kyverno",
-		Short: cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:  cobrautils.FormatDescription(false, websiteUrl, false, description...),
+		Short: command.FormatDescription(true, websiteUrl, false, description...),
+		Long:  command.FormatDescription(false, websiteUrl, false, description...),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return cmd.Help()
 		},
