+++ cmd/cli/kubectl-kyverno/commands/fix/command.go
@@ -1,17 +1,17 @@
 package fix
 
 import (
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/fix/test"
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
 	"github.com/spf13/cobra"
 )
 
 func Command() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "fix",
-		Short:   cobrautils.FormatDescription(true, websiteUrl, true, description...),
-		Long:    cobrautils.FormatDescription(false, websiteUrl, true, description...),
-		Example: cobrautils.FormatExamples(examples...),
+		Short:   command.FormatDescription(true, websiteUrl, true, description...),
+		Long:    command.FormatDescription(false, websiteUrl, true, description...),
+		Example: command.FormatExamples(examples...),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return cmd.Help()
 		},
