+++ cmd/cli/kubectl-kyverno/commands/jp/command.go
@@ -1,19 +1,19 @@
 package jp
 
 import (
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/jp/function"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/jp/parse"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/commands/jp/query"
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
 	"github.com/spf13/cobra"
 )
 
 func Command() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "jp",
-		Short:   cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:    cobrautils.FormatDescription(false, websiteUrl, false, description...),
-		Example: cobrautils.FormatExamples(examples...),
+		Short:   command.FormatDescription(true, websiteUrl, false, description...),
+		Long:    command.FormatDescription(false, websiteUrl, false, description...),
+		Example: command.FormatExamples(examples...),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return cmd.Help()
 		},
