+++ cmd/cli/kubectl-kyverno/commands/jp/function/command.go
@@ -3,7 +3,7 @@ package function
 import (
 	"fmt"
 
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/pkg/config"
 	"github.com/kyverno/kyverno/pkg/engine/jmespath"
 	"github.com/spf13/cobra"
@@ -14,9 +14,9 @@ import (
 func Command() *cobra.Command {
 	return &cobra.Command{
 		Use:          "function [function_name]...",
-		Short:        cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:         cobrautils.FormatDescription(false, websiteUrl, false, description...),
-		Example:      cobrautils.FormatExamples(examples...),
+		Short:        command.FormatDescription(true, websiteUrl, false, description...),
+		Long:         command.FormatDescription(false, websiteUrl, false, description...),
+		Example:      command.FormatExamples(examples...),
 		SilenceUsage: true,
 		Run: func(cmd *cobra.Command, args []string) {
 			printFunctions(args...)
