+++ cmd/cli/kubectl-kyverno/commands/jp/parse/command.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 
 	gojmespath "github.com/kyverno/go-jmespath"
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/spf13/cobra"
 )
 
@@ -15,9 +15,9 @@ func Command() *cobra.Command {
 	var files []string
 	cmd := &cobra.Command{
 		Use:          "parse [-f file|expression]...",
-		Short:        cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:         cobrautils.FormatDescription(false, websiteUrl, false, description...),
-		Example:      cobrautils.FormatExamples(examples...),
+		Short:        command.FormatDescription(true, websiteUrl, false, description...),
+		Long:         command.FormatDescription(false, websiteUrl, false, description...),
+		Example:      command.FormatExamples(examples...),
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			expressions, err := loadExpressions(cmd, args, files)
