+++ cmd/cli/kubectl-kyverno/commands/jp/query/command.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 
 	gojmespath "github.com/kyverno/go-jmespath"
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/pkg/config"
 	"github.com/kyverno/kyverno/pkg/engine/jmespath"
 	"github.com/spf13/cobra"
@@ -22,9 +22,9 @@ func Command() *cobra.Command {
 	var queries []string
 	cmd := &cobra.Command{
 		Use:          "query [-i input] [-q query|query]...",
-		Short:        cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:         cobrautils.FormatDescription(false, websiteUrl, false, description...),
-		Example:      cobrautils.FormatExamples(examples...),
+		Short:        command.FormatDescription(true, websiteUrl, false, description...),
+		Long:         command.FormatDescription(false, websiteUrl, false, description...),
+		Example:      command.FormatExamples(examples...),
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			queries, err := loadQueries(args, queries)
