+++ cmd/cli/kubectl-kyverno/commands/version/command.go
@@ -3,7 +3,7 @@ package version
 import (
 	"fmt"
 
-	cobrautils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/cobra"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/command"
 	"github.com/kyverno/kyverno/pkg/version"
 	"github.com/spf13/cobra"
 )
@@ -11,9 +11,9 @@ import (
 func Command() *cobra.Command {
 	return &cobra.Command{
 		Use:     "version",
-		Short:   cobrautils.FormatDescription(true, websiteUrl, false, description...),
-		Long:    cobrautils.FormatDescription(false, websiteUrl, false, description...),
-		Example: cobrautils.FormatExamples(examples...),
+		Short:   command.FormatDescription(true, websiteUrl, false, description...),
+		Long:    command.FormatDescription(false, websiteUrl, false, description...),
+		Example: command.FormatExamples(examples...),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			fmt.Printf("Version: %s\n", version.Version())
 			fmt.Printf("Time: %s\n", version.Time())
